Work on KartObjectProxy (and related)
